Add parallel implementation with OpenMP
